Reorganize engine to be more idiomatic
Removed NewEngine() in favor of Brainfuck(), and FuckFck(), TenX().
These functions now take an io.Reader as input and return an Engine
pointer and an error.  This will allow the calling context to use
anything that implements io.Reader as input, not just files on disk.

The tests have been modified to accomodate this change.

Additionally, STDIN and STDOUT handling has been improved for tests.
Instead of redirecting all of STDIN and STDOUT the Engine will use a
supplied io.Reader and io.Writer, and will fall-back on the OS's STDIN
and STDOUT if none have been specified before running Run().