Apply a bandaid to FuckFuck()
The parser really needs to be more robust.  For now, just bandaid it so
it works again.

Words read from the reader using a space delimiter will grab words
across a line as a single word: "one\ntwo".  The resulting word is split
and the slice is then looped over.

This should really be done with a proper tokenizer, but I can't be
bothered to do that right yet.