Refactor dialect handling
Dialect handling has been refactored to make adding new dialects easier.
Instead of a dedicated load function for each dialect, the dialects are
defined either as Rune based or Word based.  From here, parsing the
dialect can be done in a generic manner.  This also fixes the "bandaid"
that as previously applied in regards to newlines.

To implement a new dialect only a map of commands needs to be written
and registered.  The basic gist is each of the dialect's words or
unicode runes are mapped to a Brainfuck rune.  The resulting rune is
then loaded into the engine.

A new --language command line option has been added to force the given
dialect.  All available dialects can be listed with --dialects.

There currently isn't any support for a word based dialect that is not
space separated.  This may be added in the future.